CF refactoring: add method to configure deploy templates in resource class (#3413)
